ghostme/courseetl/requirements.txt
httpx
aiolimiter
schedule
psycopg2-binary